+++ tests/test_version.py
@@ -5,7 +5,7 @@ import xalgo
 
 class VersionTests(unittest.TestCase):
     def test_release_version(self):
-        self.assertEqual(xalgo.__version__, "0.1.3")
+        self.assertEqual(xalgo.__version__, "0.1.4")
 
 
 if __name__ == "__main__":